checks if wp exists before doenloading, current picture desc saved in wp folder

+++ gapod/gapod.py
@@ -33,14 +33,18 @@ def getNewWallpaper():
     logging.info("Find biigest file of %d images", len(imgUrls))
     apodText = getApodText(html)
     filepath = getBiggestFile(imgUrls, wallpaperPath)
+    apodText = getApodText(html)
+    save_apodText(wallpaperPath,apodText)
     logging.info("Downloaded file %s, try to set it as wallpaper", filepath)
     setNewBackground(filepath, desktopSession)
-#    sendNotification(apodText)
+    sendNotification(apodText)
+
+
 
 
 def setNewBackground(filepath, desktopSession):
     """sets Wallpaper acording to desktop session"""
-    logging.info('your WM is '+ desktopSession)
+    logging.info('your WM is ' + desktopSession)
     if 'gnome' or 'ubuntu' in desktopSession:
         os.system("gsettings set org.gnome.desktop.background picture-uri file:"+filepath)
         os.system("gsettings set org.gnome.desktop.background picture-options stretched")
@@ -48,7 +52,8 @@ def setNewBackground(filepath, desktopSession):
         os.system("feh --bg-fill "+filepath)
     if 'sway' in desktopSession:
         logging.info(' setting wallpaper in swaywm')
-        os.system("swaymsg output \"*\" background "+ filepath+' fill')
+        os.system("swaymsg output \"*\" background " + filepath+' fill')
+
 
 def getEnvironmentVars():
     """get environment vaiables from os"""
@@ -63,7 +68,7 @@ def extractImageUrlsAndHtml(date):
     logging.info(date)
     pattern = re.compile(r"image\/\d{4}\/\w*\.jpg")
     url = APODURLSTART+str(date)+APODURLEND
-    logging.warning("opening url "+url)
+    logging.info("opening url "+url)
     website = requests.get(url)
     html = website.text
     imgUrls = pattern.findall(html)
@@ -80,19 +85,21 @@ def getBiggestFile(imgUrls, wallpaperPath):
     for imgUrl in imgUrls:
         response = requests.head(APODURLSTART[:-2]+imgUrl)
         files.update({imgUrl: response.headers.get('Content-Length')})
-        logging.info(imgUrl[1]+" filesize is "+response.headers.get('Content-Length'))
+        logging.info(imgUrl+" filesize is "+response.headers.get('Content-Length'))
     biggestFile = max(files, key=lambda k: int(files[k]))
+    filepath = wallpaperPath+biggestFile[11:]
     logging.info("Biggest File is " + biggestFile[11:])
-    logging.info("starting download of "+APODURLSTART[:-2]+biggestFile)
-    image = requests.get(APODURLSTART[:-2]+biggestFile)
-    logging.info(image.status_code)
-    if image.status_code == 200:
-        filepath = wallpaperPath+biggestFile[11:]
-        logging.info("status ok, startng to write file at "+filepath)
-        with open(filepath, 'wb') as newWallpaper:
-            newWallpaper.write(image.content)
-            logging.info("file written at "+filepath)
-        del image
+    if not os.path.isfile(filepath):
+        logging.info("starting download of "+APODURLSTART[:-2]+biggestFile)
+        image = requests.get(APODURLSTART[:-2]+biggestFile)
+        if image.status_code == 200:
+            logging.info("status ok, startng to write file at "+filepath)
+            with open(filepath, 'wb') as newWallpaper:
+                newWallpaper.write(image.content)
+                logging.info("file written at "+filepath)
+            del image
+    else:
+        logging.info('file already exists at '+filepath)
     return filepath
 
 
@@ -100,8 +107,14 @@ def getApodText(html):
     """Searches and returns for the Title of the current picture"""
     pattern = re.compile(r"<b>\s(.+)\s\<\/b\>\s\<br>")
     apodText = pattern.findall(html)
-    print(apodText)
+    logging.info('short description of todays picture is ' + apodText[0])
     return apodText[0]
+
+
+def save_apodText(filepath, apodText):
+    logging.info('saving description for later use')
+    with open(filepath+'current.txt','w') as descFile:
+        descFile.write(apodText+'\n')
 
 
 def sendNotification(apodText):
@@ -124,6 +137,8 @@ def getCurrentDate():
     """Returns current date as YYMMDD string"""
     currentDate = datetime.datetime.now().strftime('%y%m%d')
     return currentDate
+
+
 
 if __name__ == '__main__':
     logging.basicConfig(level=logging.INFO)
